Add minimal configuration (auth, next servers, expire ...) Code refactor, Bugfixes

+++ api/src/main/java/me/theseems/tauth/Settings.java
@@ -13,7 +13,7 @@ public interface Settings {
      * Get mils before player's session expired
      * @return mils
      */
-    long getExpireMils();
+    Integer getExpireMils();
 
     /**
      * Get next servers
